requirements.txt
annotated-types==0.5.0
anyio==3.7.1
autopep8==2.0.4
click==8.1.8
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.103.2
flake8==5.0.4
h11==0.14.0
idna==3.10
importlib-metadata==4.2.0
mccabe==0.7.0
pycodestyle==2.10.0
pydantic==2.5.3
pydantic-core==2.14.6
pydantic-settings==2.0.3
pyflakes==2.5.0
python-dotenv==0.21.1
sniffio==1.3.1
SQLAlchemy==2.0.42
sqlmodel==0.0.24
starlette==0.27.0
tomli==2.0.1
typing-extensions==4.7.1
uvicorn==0.22.0
zipp==3.15.0
